pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "engramia"
dynamic = ["version"]
description = "Reusable execution memory and evaluation infrastructure for AI agent frameworks"
authors = [
    {name = "Marek Čermák", email = "support@engramia.dev"},
]
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE.txt"}
keywords = ["ai", "agents", "memory", "langchain", "crewai", "llm", "rag", "evaluation", "execution-memory", "busl-1.1"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.13.0",
    "numpy>=1.26",
]

[project.urls]
Homepage = "https://github.com/engramia/engramia"
Documentation = "https://engramia.readthedocs.io"
Repository = "https://github.com/engramia/engramia"
"Bug Tracker" = "https://github.com/engramia/engramia/issues"
Changelog = "https://github.com/engramia/engramia/blob/main/CHANGELOG.md"

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.50"]
local = ["sentence-transformers>=5.4.0"]
postgres = [
    "sqlalchemy[asyncio]>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "alembic>=1.18.4",
]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.44.0",
]
cli = [
    "typer[all]>=0.12",
    "rich>=13.0",
]
langchain = ["langchain-core>=1.2.28"]
crewai = ["crewai>=0.28"]
openai-agents = ["openai-agents>=0.13.6"]
anthropic-agents = ["claude-agent-sdk>=0.1"]
pydantic-ai = ["pydantic-ai>=1.80.0"]
autogen = ["autogen-agentchat>=0.4"]
mcp = ["mcp>=1.0"]
oidc = ["PyJWT>=2.8", "cryptography>=46.0.7"]
cloud-auth = [
    "bcrypt>=5.0.0",
    "PyJWT>=2.8",
    "cryptography>=42.0",
]
billing = ["stripe>=15.0.1"]
telemetry = [
    "opentelemetry-api>=1.41.0",
    "opentelemetry-sdk>=1.41.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.41.0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "prometheus-client>=0.20",
    "python-json-logger>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.9",
    "mypy>=1.14",
    "testcontainers[postgres]>=4.0",
    # Transitive deps pinned for CVE remediation (CVE-2026-4539).
    "pygments>=2.20.0",
]
all = [
    "engramia[openai,anthropic,local,postgres,api,cli,langchain,crewai,mcp,telemetry,cloud-auth]",
]

[project.scripts]
engramia = "engramia.cli.main:app"
engramia-mcp = "engramia.mcp.server:main"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["engramia"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "--cov=engramia --cov-fail-under=80"
markers = [
    "postgres: integration tests requiring Docker + PostgreSQL+pgvector container",
    "local_embeddings: tests requiring sentence-transformers (pip install engramia[local])",
    "concurrent: thread-safety stress tests (flaky on CI tmpfs)",
    "integration: tests using create_app() with full middleware stack",
    "security: security regression tests (RBAC, tenant isolation, auth bypass)",
    "contract: interface and schema compliance tests",
    "e2e: end-to-end multi-step business flow tests",
]

[tool.coverage.run]
source = ["engramia"]
omit = ["engramia/db/migrations/*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
    "SIM108", # use ternary operator — often less readable
]

[tool.ruff.lint.isort]
known-first-party = ["engramia"]

[tool.bandit]
exclude_dirs = ["tests", "engramia/db/migrations"]
skips = ["B101"]  # assert used in tests and development guards

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
